Stop tyring to use vagrant as user for api
This is no longer necessary, because we are not writing out these logs
any longer and does not work as expected.

Also, $UID is not actually set, because it is not an envvar.

This was just defaulting to `root` - not the current UID because the string
became `":"`, which defaults to root.